village360-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
bcrypt
python-jose
alembic
aiomysql
python-dotenv